+++ EntraAuth/internal/classes/01-FederationProvider.ps1
@@ -1,0 +1,13 @@
+﻿class FederationProvider {
+	[string]$Name
+	[string]$Description
+	[int]$Priority
+	[scriptblock]$Test
+	[scriptblock]$Code
+	[string]$Assertion
+	[string]$Type = 'Custom'
+
+	[string]ToString() {
+		return $this.Name
+	}
+}