+++ EntraAuth/internal/scripts/01-variables.ps1
@@ -4,5 +4,8 @@ $script:_EntraTokens = @{}
 # Endpoint Configuration for Requests
 $script:_EntraEndpoints = @{}
 
+# Plugins for automatic processing of Federated Authentication in the correct context
+$script:_FederationProviders = @{}
+
 # The default service to connect to
 $script:_DefaultService = 'Graph'